Refactor flamechart specific logic out of SearchView
